requirements.txt
Django==1.6.3
South==0.8.4
argparse
psycopg2==2.5.2
wsgiref==0.1.2
microdata==0.4.1
fabric
uwsgi
django_extensions
django-csvimport
easy_thumbnails
pillow
pylibmc
FeedParser
hg+https://bitbucket.org/izi/django-admin-tools
git+git://github.com/ronnix/fabtools.git
python-social-auth==0.1.23
pygraphviz
pyparsing==1.5.7
pydot
BeautifulSoup4
html5lib
djangorestframework==2.3.13
requests
requesocks
factory_boy
coverage
celery
django-celery
raven
selenium
pytils
pyvirtualdisplay
gdata
zerorpc
lxml
pycurl
django-treebeard
djorm-ext-pgfulltext
GitPython
PySocks
django-nose
ipdb

#sudo apt-get install -y mercurial memcached libmemcached-dev zlib1g-dev libssl-dev python-dev build-essential libjpeg-dev libfreetype6-dev zlib1g-dev libpng12-dev libpq-dev libgraphviz-dev graphviz nginx python-virtualenv python-pygraphviz pkg-config postgresql nodejs git-flow libxslt-dev libxml2-dev libcurl4-openssl-dev curl tor ipdb libzmq-dev
#sudo npm install jade zerorpc
